feat(pipeline/replica): optimize network segment calculation for chunked inputs

+++ data-pipeline/src/etl/sources/replica/transformers/to_vector_tiles.py
@@ -5,9 +5,15 @@ import subprocess
 from typing import Any, Generator
 
 import geopandas
+import pyogrio
 
 
-def to_vector_tiles(gdf: geopandas.GeoDataFrame, name: str, layer_name: str, output_folder: str, zoomLevel: int = -1) -> Generator[float, None, None]:
+class NoVectorDataError(ValueError):
+    """Custom exception raised when no vector data is found in the input."""
+    pass
+
+
+def to_vector_tiles(gdf: geopandas.GeoDataFrame | str, name: str, layer_name: str, output_folder: str, zoomLevel: int = -1) -> Generator[float, None, None]:
     """Converts a GeoDataFrame of lines to vector tiles using tippecanoe.
 
     This function saves the GeoDataFrame to a GeoJSON file, then uses tippecanoe
@@ -20,28 +26,53 @@ def to_vector_tiles(gdf: geopandas.GeoDataFrame, name: str, layer_name: str, out
     for information about using tiles from tippecanoe in ArcGIS SDK for JavaScript clients.
 
     Args:
-        gdf (geopandas.GeoDataFrame): _description_
+        gdf (geopandas.GeoDataFrame | str): A GeoDataFrame containing line geometries or a path to a or FlatGeobuf file with the EPSG:3857 (web mercator) CRS.
         name (str): The name of the vector tile layer.
         layer_name (str): The name of the layer in the vector tiles. This name will be used to identify the layer during styling.
         output_folder (Path): The output location for the vector tiles.
         zoomLevel (int): A number from 0 to 22 indicating the zoom level for the vector tiles. Indicate -1 for auto detection. See https://github.com/felt/tippecanoe?tab=readme-ov-file#zoom-levels
     """
+    input_fgb_path = None
+    should_delete_temp_file = False
 
-    # require line geometry
-    # if not gdf.geometry.isin(['LineString', 'MultiLineString']).all():
-    #     print(
-    #         f'Warning: GeoDataFrame must contain only LineString or MultiLineString geometries. Found: {gdf.geometry.unique()}')
+    if isinstance(gdf, str):
+        # require .geojson or .fgb file
+        if not gdf.endswith(('.geojson', '.fgb')):
+            raise ValueError(
+                f"Input file must be a GeoJSON file or a FlatGeobuf file. Found: {gdf}")
 
-    # reproject to Web Mercator
-    temp_geojson_path = os.path.join(output_folder, f'{layer_name}.geojson')
+        # if gdf is a string, assume it's a path to a GeoDataFrame file
+        if not os.path.exists(gdf):
+            raise FileNotFoundError(f"File not found: {gdf}")
 
-    # ensure the output folder exists and is empty
-    if os.path.exists(output_folder):
-        shutil.rmtree(output_folder)
-    os.makedirs(output_folder, exist_ok=True)
+        # ensure that the CRS is EPSG:3857 (web mercator)
+        # without reading the whole file into memory
+        info = pyogrio.read_info(gdf)
+        crs = info['crs']
+        if (crs is None or crs != 'EPSG:3857'):
+            raise ValueError(
+                f"GeoDataFrame must be in EPSG:3857 (web mercator) CRS. Found: {crs}")
 
-    # save to file so we can use it with tippecanoe
-    gdf.to_crs('EPSG:3857').to_file(temp_geojson_path, driver='GeoJSON')
+        # skip if the input file has no features (note that this is not always computed)
+        if info['features'] == 0:
+            raise NoVectorDataError(f"GeoDataFrame file {gdf} has no features.")
+
+        input_fgb_path = gdf
+
+    else:
+        input_fgb_path = os.path.join(output_folder, f'{layer_name}.fgb')
+        should_delete_temp_file = True
+
+        if len(gdf) == 0:
+            raise NoVectorDataError(f"GeoDataFrame has no features.")
+
+        # ensure the output folder exists and is empty
+        if os.path.exists(output_folder):
+            shutil.rmtree(output_folder)
+        os.makedirs(output_folder, exist_ok=True)
+
+        # reproject to Web Mercator and save to file so we can use it with tippecanoe
+        gdf.to_crs('EPSG:3857').to_file(input_fgb_path, driver='FlatGeobuf')
 
     # generate vector tiles using tippecanoe - see https://github.com/felt/tippecanoe
     command = [
@@ -55,11 +86,15 @@ def to_vector_tiles(gdf: geopandas.GeoDataFrame, name: str, layer_name: str, out
         # dynamically drop features at a zoom level if a tile at that zoom level is too large (> 500 KB)
         '--drop-fraction-as-needed',
         '--json-progress',
-        temp_geojson_path
+        input_fgb_path
     ]
 
-    process = subprocess.Popen(command, stdout=subprocess.PIPE,
-                               stderr=subprocess.STDOUT, text=True)
+    process = subprocess.Popen(
+        command,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        text=True
+    )
 
     if process.stdout is None:
         raise RuntimeError(
@@ -74,7 +109,12 @@ def to_vector_tiles(gdf: geopandas.GeoDataFrame, name: str, layer_name: str, out
                 yield progress_percent
 
         except json.JSONDecodeError:
+            if (line.strip() == 'Did not read any valid geometries'):
+                raise NoVectorDataError(
+                    "No valid geometries found in the input file. Please check the input data.")
+
             # ignore lines that are not valid JSON
+            print(f"\nWarning: Skipping non-JSON line: {line.strip()}")
             pass
 
         except Exception as e:
@@ -136,8 +176,8 @@ def to_vector_tiles(gdf: geopandas.GeoDataFrame, name: str, layer_name: str, out
         json.dump(style, file, indent=2)
 
     # delete the generated GeoJSON file
-    if os.path.exists(temp_geojson_path):
-        os.remove(temp_geojson_path)
+    if should_delete_temp_file and os.path.exists(input_fgb_path):
+        os.remove(input_fgb_path)
 
 
 def create_vector_tile_server_index(name: str) -> dict[str, Any]:
